Moved merkle_sign, merkle_gen_root into merkle.c
